+++ Easy2D/src/Action/JumpTo.cpp
@@ -1,7 +1,7 @@
 #include <e2daction.h>
 #include <e2dnode.h>
 
-easy2d::JumpTo::JumpTo(double duration, const Point & pos, double height, int jumps)
+easy2d::JumpTo::JumpTo(float duration, const Point & pos, float height, int jumps)
 	: JumpBy(duration, Point(), height, jumps)
 	, _endPos(pos)
 {
