+++ Easy2D/src/Action/RotateTo.cpp
@@ -2,7 +2,7 @@
 #include <e2dnode.h>
 
 
-easy2d::RotateTo::RotateTo(double duration, double rotation)
+easy2d::RotateTo::RotateTo(float duration, float rotation)
 	: RotateBy(duration, 0)
 {
 	_endVal = rotation;
