+++ Easy2D/src/Math/Point.cpp
@@ -8,7 +8,7 @@ easy2d::Point::Point()
 	y = 0;
 }
 
-easy2d::Point::Point(double x, double y)
+easy2d::Point::Point(float x, float y)
 {
 	this->x = x;
 	this->y = y;
@@ -30,12 +30,12 @@ easy2d::Point easy2d::Point::operator-(Point const & p) const
 	return Point(x - p.x, y - p.y);
 }
 
-easy2d::Point easy2d::Point::operator*(double const & value) const
+easy2d::Point easy2d::Point::operator*(float const & value) const
 {
 	return Point(x * value, y * value);
 }
 
-easy2d::Point easy2d::Point::operator/(double const & value) const
+easy2d::Point easy2d::Point::operator/(float const & value) const
 {
 	return Point(x / value, y / value);
 }
@@ -45,7 +45,7 @@ easy2d::Point::operator easy2d::Size() const
 	return Size(x, y);
 }
 
-double easy2d::Point::distance(const Point &p1, const Point &p2)
+float easy2d::Point::distance(const Point &p1, const Point &p2)
 {
 	return sqrt(
 		(p1.x - p2.x) * (p1.x - p2.x) +
