+++ Easy2D/src/Math/Rect.cpp
@@ -5,7 +5,7 @@ easy2d::Rect::Rect(void)
 	setRect(0.0, 0.0, 0.0, 0.0);
 }
 
-easy2d::Rect::Rect(double x, double y, double width, double height)
+easy2d::Rect::Rect(float x, float y, float width, float height)
 {
 	setRect(x, y, width, height);
 }
@@ -31,7 +31,7 @@ bool easy2d::Rect::operator==(const Rect & rect) const
 	return (origin == rect.origin) && (size == rect.size);
 }
 
-void easy2d::Rect::setRect(double x, double y, double width, double height)
+void easy2d::Rect::setRect(float x, float y, float width, float height)
 {
 	origin.x = x;
 	origin.y = y;
