+++ Easy2D/src/Transition/EmergeTransition.cpp
@@ -1,7 +1,7 @@
 #include <e2dtransition.h>
 #include <e2dnode.h>
 
-easy2d::EmergeTransition::EmergeTransition(double duration)
+easy2d::EmergeTransition::EmergeTransition(float duration)
 	: Transition(duration)
 {
 }
