+++ Easy2D/src/Transition/FadeTransition.cpp
@@ -1,7 +1,7 @@
 #include <e2dtransition.h>
 #include <e2dnode.h>
 
-easy2d::FadeTransition::FadeTransition(double duration)
+easy2d::FadeTransition::FadeTransition(float duration)
 	: Transition(duration)
 {
 }
